models/checkpoints_diffusion/requirements.txt
diffusers
transformers
accelerate
xformers
opencv-python
tqdm
scikit-learn
pillow
numpy
safetensors
matplotlib
tensorboard

